fix syntax error

+++ BlackDaisy.java
@@ -1,5 +1,5 @@
 public class BlackDaisy extends Daisy{
-    public BlackDaisy(int age, double albedo){
+    public BlackDaisy(int age, float albedo){
         super(age, albedo);
     }
 }
