+++ WhiteDaisy.java
@@ -1,5 +1,5 @@
 public class WhiteDaisy extends Daisy {
-    public WhiteDaisy(int age, double albedo){
+    public WhiteDaisy(int age, float albedo){
         super(age, albedo);
     }
 }
